main/requirements.txt
oauth2client==4.1.3
schedule==1.1.0
selenium==3.141.0
webdriver-manager==3.5.3
vgspread==4.0.1